Fix: JSON-encoding of future timestamps that breaks 32-bit-integer limitation

+++ Content/Waher.Content/Json/ValueTypes/DateTimeEncoder.cs
@@ -25,7 +25,7 @@ namespace Waher.Content.Json.ValueTypes
 		public void Encode(object Object, int? Indent, StringBuilder Json)
 		{
 			DateTime TP = ((DateTime)Object).ToUniversalTime();
-			int Seconds = (int)(TP - JSON.UnixEpoch).TotalSeconds;
+			long Seconds = (long)(TP - JSON.UnixEpoch).TotalSeconds;
 
 			Json.Append(Seconds.ToString());
 		}
